[refactor] 마이페이지 탭별로 뷰모델 분리, Content, Component 파일로 분리
