+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/mypage/MyPageRoute.kt
@@ -5,6 +5,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.paging.compose.collectAsLazyPagingItems
+import com.dpm.sixpack.presentation.routes.mypage.contract.MyPageFeedTabSideEffect
+import com.dpm.sixpack.presentation.routes.mypage.contract.MyPageRecordTabSideEffect
 import com.dpm.sixpack.presentation.routes.mypage.contract.MyPageSideEffect
 import com.dpm.sixpack.presentation.routes.mypage.ui.screen.MyPageScreen
 import org.orbitmvi.orbit.compose.collectAsState
@@ -14,23 +16,44 @@ import org.orbitmvi.orbit.compose.collectSideEffect
 fun MyPageRoute(
     modifier: Modifier = Modifier,
     viewModel: MyPageViewModel = hiltViewModel(),
+    feedTabViewModel: MyPageFeedTabViewModel = hiltViewModel(),
+    recordTabViewModel: MyPageRecordTabViewModel = hiltViewModel(),
     onNavigateToSettings: () -> Unit = {},
+    onNavigateToPostDetail: (Long) -> Unit = {},
     onNavigateToRecordDetail: (Long) -> Unit = {},
 ) {
     val state by viewModel.collectAsState()
-    val gridItemsPagingItems = viewModel.postsPagingFlow.collectAsLazyPagingItems()
+    val feedTabState by feedTabViewModel.collectAsState()
+    val recordTabState by recordTabViewModel.collectAsState()
+
+    val gridItemsPagingItems = feedTabViewModel.postsPagingFlow.collectAsLazyPagingItems()
 
     viewModel.collectSideEffect { sideEffect ->
         when (sideEffect) {
             is MyPageSideEffect.NavigateToSettings -> onNavigateToSettings()
-            is MyPageSideEffect.NavigateToRecordDetail -> onNavigateToRecordDetail(sideEffect.recordId)
+        }
+    }
+
+    feedTabViewModel.collectSideEffect { sideEffect ->
+        when (sideEffect) {
+            is MyPageFeedTabSideEffect.NavigateToPostDetail -> onNavigateToPostDetail(sideEffect.postId)
+        }
+    }
+
+    recordTabViewModel.collectSideEffect { sideEffect ->
+        when (sideEffect) {
+            is MyPageRecordTabSideEffect.NavigateToRecordDetail -> onNavigateToRecordDetail(sideEffect.recordId)
         }
     }
 
     MyPageScreen(
         state = state,
+        feedTabState = feedTabState,
+        recordTabState = recordTabState,
         gridItemsPagingItems = gridItemsPagingItems,
         onIntent = viewModel::onIntent,
+        onFeedTabIntent = feedTabViewModel::onIntent,
+        onRecordTabIntent = recordTabViewModel::onIntent,
         modifier = modifier,
     )
 }
