+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/mypage/contract/MyPageRecordTabIntent.kt
@@ -1,0 +1,13 @@
+package com.dpm.sixpack.presentation.routes.mypage.contract
+
+import com.dpm.sixpack.presentation.common.base.UiIntent
+
+sealed interface MyPageRecordTabIntent : UiIntent {
+    data object OnPreviousMonthClick : MyPageRecordTabIntent
+
+    data object OnNextMonthClick : MyPageRecordTabIntent
+
+    data class OnRecordClick(
+        val recordId: Long,
+    ) : MyPageRecordTabIntent
+}
