+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/mypage/contract/MyPageRecordTabSideEffect.kt
@@ -1,0 +1,9 @@
+package com.dpm.sixpack.presentation.routes.mypage.contract
+
+import com.dpm.sixpack.presentation.common.base.SideEffect
+
+sealed interface MyPageRecordTabSideEffect : SideEffect {
+    data class NavigateToRecordDetail(
+        val recordId: Long,
+    ) : MyPageRecordTabSideEffect
+}
