+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/mypage/contract/MyPageSideEffect.kt
@@ -4,8 +4,4 @@ import com.dpm.sixpack.presentation.common.base.SideEffect
 
 sealed interface MyPageSideEffect : SideEffect {
     data object NavigateToSettings : MyPageSideEffect
-
-    data class NavigateToRecordDetail(
-        val recordId: Long,
-    ) : MyPageSideEffect
 }
